LaTeX @def* in code, upright brackets and parentheses
* tp/Texinfo/Convert/LaTeX.pm (_protect_text): do not protect
brackets.  It is needed for embrac package to work, but it
is also useless to protect brackets as we already protect \
such that there cannot be a situation where [ ] can be interpreted
as macro optional argument.
* tp/Texinfo/Convert/LaTeX.pm (_latex_header, _stop_embrac)
(_restart_embrac_if_needed, _convert): use the embrac package
to format upright parentheses and brackets on @def* line.  Turn
off embrac when opening a style commands.
remove _only_slanted_no_code_contents and associated types.