Avoid alignment attributes in HTML when not needed
* tp/Texinfo/Convert/HTML.pm: remove valign="middle" and
align="left" from <td> as it is the default, except for @def*
in table in which the align left is important.  Remove align=
from tables with only one line, and valign= from tables with
only one column.  Remove valign="top" from 'Jump to' in
printindex summary letter as it is better without.