@deftype* as generic commands, @deffn with simple arguments
* doc/texinfo.texi (Definition Commands, Object-Oriented Methods):
describe @deffn and other @-command with the whole arguments
formatted as @var as having simple arguments.
(Typed Functions, Object-Oriented Methods): change wording to make
clearer that the @deftypefn and @deftypeop are generic command,
not only suitable for typed languages.
(Typed Functions): remove a confusing precision for the Ada example.
Add a Perl example with named variables as return value information.

* doc/texinfo.texi (Def Cmd Conventions): remove 'with the following
word', as the word is already marked.