* doc/texinfo.texi (Node Line Requirements): add in comment that spaces from @-commands are collapsed too.
* doc/texinfo.texi (Node Line Requirements, Menu Parts): add
a FIXME in comment to remove the constraint on extra space in
Info as both the standalone and Emacs Info reader handle multiple
spaces right now.

* doc/texinfo.texi (HTML Xref Node Name Expansion): newlines in
node names are correctly handled in all the output formats, remove
the parenthesis stating that it is not the case.

* doc/texinfo.texi (Info Format Menu): add a precision that in menu
there is * instead of * (N|n)ote in cross-references.