Hypenation in @code
* doc/texinfo.tex (\code, \codedash): Set internal \codedashprev
macro differently when \code starts to avoid hyphenation being
always turned off for the first hyphen in code.  \codedashprev
does not actually refer to the previous character, as it is
only set when a dash is seen, not at every single character
within @code, so rename it to \codedashnobreak to be less
confusing.

Report from Werner.