* tp/Texinfo/Convert/Text.pm: rename _code_options as _code_state and raw as _raw_state.
* tp/Texinfo/Convert/Text.pm (_convert): decrease _raw_state when
coming out of raw state.

* tp/Texinfo/ParserNonXS.pm (_process_remaining_on_line): in ignored
@inline* arguments, do not postpend to the line when there is no match
with a delimiter, instead consume the line. As in the XS parser.

* tp/Texinfo/XS/parsetexi/macro.c (expand_macro_body): keep text after
a non paired \ in the macro body, like the perl parser.

* tp/Texinfo/ParserNonXS.pm (_process_remaining_on_line),
tp/Texinfo/XS/parsetexi/separator.c (handle_comma): put elided
content of @inline* in the tree.

* tp/Texinfo/ParserNonXS.pm (_expand_macro_body): change code to be
clearer and more in line with XS parser code.