* tp/Texinfo/ParserNonXS.pm (_end_line_misc_line) (_end_line_starting_block, _check_register_target_element_label) (_process_remaining_on_line), tp/Texinfo/XS/parsetexi/end_line.c (check_register_target_element_label, end_line_starting_block) (end_line_misc_line), tp/Texinfo/XS/parsetexi/separator.c (handle_close_brace): group together the call of parse_node_manual, the check of no manual element and the registering if a target element for links in _check_register_target_element_label.  Remove _check_internal_node, integrated in _check_register_target_element_label.
* tp/Texinfo/XS/parsetexi/end_line.c (parse_node_manual): add an
argument to control whether the element argument is modified or not.
Add a field in NODE_SPEC_EXTRA to keep the added elements in case
the element argument is not modified.  Do not modify the node before
being sure that there really is a manual name if there is an opening
brace.  Update callers.