TOC section number alignment
* doc/texinfo.tex (\findsecnowidths): New macro, to get maximum
width of section numbers for each chapter.
(\contents): Call it.

(\retrievesecnowidths): New macro to get the maximum widths.
(\numchapentry, \appentry): Call it.

(\tocentry, \secnowidth): Use \secnowdith for alignment if set to
a non-zero value.
(\dosecentry): Set \secnowidth based on result of \retrievesecnowidths.

This aligns section titles in the table of contents if there are
10 or more sections in a chapter.  Report from Karl.