Revert documentlanguage fix
* tp/Texinfo/ParserNonXS.pm (_parse_def),
* tp/Texinfo/XS/parsetexi/def.c (parse_def): Do not set type
untranslated if @documentlanguage is not set at the time of parsing.
* tp/t/languages.t (no_documentlanguage_before_copying): Skip test.

This reverses change on 2023-01-11.  Although it is correct,
it slows the program down a lot when a lot of @def* are used
and there is no @documentlanguage.

This gets back to the execution times from texi2any 7.1 (with XS
enabled only).