No encoding, no us-ascii added to locale for document translations
* NEWS, doc/texinfo.texi (Internationalization of Document Strings),
tp/Texinfo/Translations.pm (translate_string),
tp/Texinfo/XS/main/translations.c (translate_string): do not prepend
the encoding nor us-ascii to the locale name to find the translations.
For the encoding it is not done in XS, not useful since long in Perl
and encoding was never available when translating from parser.  Note
that this does not constrain the actual encoding declared in the po/mo
file (which could be us-ascii) and do not prevent from using accented
@-commands in translations (which would be needed for accented letters
if the encoding is us-ascii).