Only allow highlighting commands to be redefined with @definfoenclose
* tp/Texinfo/ParserNonXS.pm (_parse_line_command_args),
tp/Texinfo/XS/parsetexi/end_line.c (parse_line_command_args):
before redefining with definfoenclose, verify that the command is not
a builtin command, or, if it is, that it is a brace command and a
style* brace command.

* tp/Makefile.tres, tp/t/03coverage_braces.t
(definfoenclose_texinfo_commands): add tests of redefinition of
builtin @-commands by @definfoenclose.