Fix error message of XS parser when the input file cannot be opened
* tp/Texinfo/XS/parsetexi/api.c (initialize_parsing, parse_file)
(parse_text, parse_string, parse_piece): have initialize_parsing
return the document_descriptor and get it from there in parse_*.

* tp/Texinfo/XS/parsetexi/Parsetexi.pm (parse_texi_file),
tp/Texinfo/XS/parsetexi/Parsetexi.xs (parse_file),
tp/Texinfo/XS/parsetexi/api.c (parse_file): add a status
argument to parse_file in api.c, set it to non-zero if the parsing
could not proceed.  Set the open error message in api.c parse_file,
where the error can be converted to a string with strerror.  Return a
document_descriptor even if there was an error.  In Parsetexi.xs
parse_file, check the status, and if there was an error, get the
parsing error and destroy the document.  Remove the error setting
in Parsetexi.pm parse_texi_file, where $! is unlikely to be correct.

* tp/Texinfo/ParserNonXS.pm (parse_texi_file): rename $input_file_name
as $decoded_input_file_path.

* tp/Texinfo/ParserNonXS.pm (parse_texi_file): do not pass the
PROGRAM configuration value to document_error, it is not in the parser
customization variables, and we do not want to add more.