* info/echo-area.c (remove_completion_duplicates, build_completions): get completions number in argument and return the new completions number.
* info/echo-area.c (build_completions): use a size_t index variable to
add entries to completions_found.  Pass this variable to
remove_completion_duplicates and retirieve the modified value.  Set
completions_found_index at this point, such that the conversion to
long is explicit, and mark in a comment that the completions_found*
and LCD_completion should not be modified after that point and before
a call to build_completions.